+++ tests/inative.rs
@@ -47,6 +47,12 @@ fn try_from() {
         Ok(INative::from(42i8))
     );
     assert!(INative::try_from(UNative::MAX).is_err());
+
+    // Conversion to `bool`.
+    assert_eq!(bool::try_from(INative::ZERO), Ok(false));
+    assert_eq!(bool::try_from(INative::from(1i8)), Ok(true));
+    assert!(bool::try_from(INative::from(2i8)).is_err());
+    assert!(bool::try_from(INative::from(-1i8)).is_err());
 }
 
 #[test]